control_volumen_manos_DZF/requirements.txt
pymongo[srv]
python-dotenv
opencv-python
mediapipe
pycaw
comtypes
